move actors to helpers_test
